+++ packages/diagrams/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @myst-theme/diagrams
 
+## 1.1.2
+
+### Patch Changes
+
+- @myst-theme/providers@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- @myst-theme/providers@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- Updated dependencies [fd4708d]
+  - @myst-theme/providers@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes