+++ packages/providers/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @myst-theme/providers
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [c3c427d]
+  - @myst-theme/common@1.1.2
+  - @myst-theme/search@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- @myst-theme/common@1.1.1
+- @myst-theme/search@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- fd4708d: Export myst-search type dependence of providers
+  - @myst-theme/common@1.1.0
+  - @myst-theme/search@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes