+++ packages/search-minisearch/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @myst-theme/search-minisearch
 
+## 1.1.2
+
+### Patch Changes
+
+- @myst-theme/search@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- @myst-theme/search@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- @myst-theme/search@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes