+++ packages/site/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @myst-theme/site
 
+## 1.1.2
+
+### Patch Changes
+
+- a569399: Make sidebar modal take up full height
+- 2208b1e: Re-use classes for sidebar whitespace
+- 192a1b0: Use parent width instead of screen size for page components to fix overflow
+- c3c427d: Add alt text for logo
+- Updated dependencies [7d92ceb]
+- Updated dependencies [c3c427d]
+  - myst-to-react@1.1.2
+  - @myst-theme/common@1.1.2
+  - @myst-theme/providers@1.1.2
+  - @myst-theme/frontmatter@1.1.2
+  - @myst-theme/diagrams@1.1.2
+  - @myst-theme/search@1.1.2
+  - myst-demo@1.1.2
+
+## 1.1.1
+
+### Patch Changes
+
+- f722531: Increase contrast of search bar on dark mode
+- Updated dependencies [ae018ba]
+  - myst-to-react@1.1.1
+  - @myst-theme/providers@1.1.1
+  - @myst-theme/frontmatter@1.1.1
+  - @myst-theme/diagrams@1.1.1
+  - @myst-theme/common@1.1.1
+  - @myst-theme/search@1.1.1
+  - myst-demo@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- e213587: On static builds, fix sidebar menu not expanding for current page (closes [#635](https://github.com/jupyter-book/mystmd/issues/2288)).
+- 37fa906: Improve UX of ToC expand widget
+- Updated dependencies [fd4708d]
+- Updated dependencies [75475f8]
+- Updated dependencies [c519f9b]
+- Updated dependencies [9bfb08e]
+  - @myst-theme/providers@1.1.0
+  - myst-to-react@1.1.0
+  - @myst-theme/frontmatter@1.1.0
+  - @myst-theme/diagrams@1.1.0
+  - @myst-theme/common@1.1.0
+  - @myst-theme/search@1.1.0
+  - myst-demo@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes